This project aims at developing Android application to find nearby places and related comments for the users.
This project make use of Android SDK and Google Play Service to display
Google map and pinpoint the user on the map. A list of nearby
restaurants or other places are pinpointed on the Google Map. Users can
upload their comments after logging in. The mobile phone front end is
connected with database by PHP.